+++ ModelTracker/models.py
@@ -1,4 +1,12 @@
+import datetime
+from functools import lru_cache
+
 from django.db import models
+from django.apps import apps
+
+from ModelTracker.utils import get_fields,get_model
+
+
 try:
     from django.db.models import JSONField
 except ImportError:
@@ -6,7 +14,6 @@ except ImportError:
         from jsonfield.fields import JSONField
     except ImportError:
         raise ImportError("Can't find a JSONField implementation, please install jsonfield if django < 4.0")
-
 
 
 class History(models.Model):
@@ -19,8 +26,26 @@ class History(models.Model):
     done_by=models.CharField(max_length=255)
     done_on=models.DateTimeField(auto_now_add=True)
 
+    def get_object(self):
+        model = get_model(self.table)
+        fields = get_fields(model)
+        keys2del = []
+        state = self.new_state
+        for key in state:
+            if (key.startswith("_") and "_cache" in key) or (key not in fields and not ("_id" in key and key[:-3] in fields)):
+                keys2del.append(key)
+            if type(state[key])==type({}):
+                if state[key].get("_type",None) == "datetime":
+                    state[key]=datetime.datetime.strptime(state[key]["value"],"%Y-%m-%d %H:%M:%S")
+                elif state[key].get("_type",None) == "date":
+                    state[key]=datetime.datetime.strptime(state[key]["value"],"%Y-%m-%d")
+        for key in keys2del:
+            del state[key]
+        return model(**state)
+
     def __eq__(self, other):
         return self.id == other.id
+
     def __str__(self):
         if self.name:
             return "%s for %s in %s"%(self.name, self.primary_key, self.table)